continue CFIHOS_requirements plugin
