+++ public/vocables/index.html
@@ -163,7 +163,7 @@
 
         <script type="module" src="modules/tools/lineage/lineage_similars.js"></script>
 
-        <script type="module" src="modules/tools/ontoWagon/ontoWagon.js"></script>
+        <script type="module" src="modules/tools/weaver/weaver.js"></script>
 
         <script type="module" src="modules/shared/graphLoader.js"></script>
